% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes-cube.R
\docType{data}
\name{aggregate_spatial}
\alias{aggregate_spatial}
\title{aggregate temporal period}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
aggregate_spatial
}
\description{
aggregate temporal period
}
\keyword{datasets}
