% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{aggregate_temporal_period}
\alias{aggregate_temporal_period}
\title{aggregate temporal period}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
aggregate_temporal_period
}
\description{
aggregate temporal period
}
\keyword{datasets}
