% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter-class.R
\name{appendInfo}
\alias{appendInfo}
\title{appendInfo}
\usage{
appendInfo(name, description, schema, optional = NA)
}
\arguments{
\item{name}{name of parameter}

\item{description}{description of parameter}

\item{type}{type of parameter}

\item{subtype}{subtype of parameter}
}
\value{
list with appended parameter
}
\description{
Create a list with appended parameter
}
