% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_prediction.R
\docType{data}
\name{apply_prediction}
\alias{apply_prediction}
\title{apply_prediction}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
apply_prediction
}
\description{
apply_prediction
}
\keyword{datasets}
