% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{array_element}
\alias{array_element}
\title{array element}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
array_element
}
\description{
array element
}
\keyword{datasets}
