% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{array_interpolate_linear}
\alias{array_interpolate_linear}
\title{array interpolate linear}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
array_interpolate_linear
}
\description{
array interpolate linear
}
\keyword{datasets}
