% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{changeProjection}
\alias{changeProjection}
\title{changeProjection}
\usage{
changeProjection(extent)
}
\arguments{
\item{extent}{bbox of spatial extent}
}
\value{
changed spatial extent
}
\description{
Change from EPSG:4326 to EPSG:3857
}
