% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session-Class.R
\name{createSessionInstance}
\alias{createSessionInstance}
\title{Creates a new instance from the class 'SessionInstance' and assigns the name 'Session'}
\usage{
createSessionInstance(configuration = NULL)
}
\arguments{
\item{configuration}{Edited configuration for the session}
}
\description{
Creates a new instance from the class 'SessionInstance' and assigns the name 'Session'
}
