% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{eo_datacube}
\alias{eo_datacube}
\title{return object for the processes}
\format{
An object of class \code{list} of length 2.
}
\usage{
eo_datacube
}
\description{
return object for the processes
}
\keyword{datasets}
