% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutableProcess-class.R
\name{ExecutableProcess}
\alias{ExecutableProcess}
\title{Executable process}
\description{
Executable process

Executable process
}
\section{Super class}{
\code{\link[openeocubes:Process]{openeocubes::Process}} -> \code{ExecutableProcess}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{job}}{Attached job to the process}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExecutableProcess-new}{\code{ExecutableProcess$new()}}
\item \href{#method-ExecutableProcess-execute}{\code{ExecutableProcess$execute()}}
\item \href{#method-ExecutableProcess-clone}{\code{ExecutableProcess$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="openeocubes" data-topic="Process" data-id="processInfo"><a href='../../openeocubes/html/Process.html#method-Process-processInfo'><code>openeocubes::Process$processInfo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="openeocubes" data-topic="Process" data-id="setParameter"><a href='../../openeocubes/html/Process.html#method-Process-setParameter'><code>openeocubes::Process$setParameter()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutableProcess-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutableProcess-new}{}}}
\subsection{Method \code{new()}}{
Initialize executable process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutableProcess$new(
  id = NA,
  description = NA,
  parameters = NA,
  operation = NA,
  process = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id or name of the process}

\item{\code{description}}{Shortly what the process does}

\item{\code{parameters}}{Used parameters in the process}

\item{\code{operation}}{Function that executes the process}

\item{\code{process}}{Processes which will be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutableProcess-execute"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutableProcess-execute}{}}}
\subsection{Method \code{execute()}}{
Run the operation including a generated list of parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutableProcess$execute()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutableProcess-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutableProcess-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutableProcess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
