% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{filter_bands}
\alias{filter_bands}
\title{filter bands}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
filter_bands
}
\description{
filter bands
}
\keyword{datasets}
