% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{filter_bbox}
\alias{filter_bbox}
\title{filter bbox}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
filter_bbox
}
\description{
filter bbox
}
\keyword{datasets}
