% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{filter_spatial}
\alias{filter_spatial}
\title{filter_spatial}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
filter_spatial
}
\description{
filter_spatial
}
\keyword{datasets}
