% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{filter_temporal}
\alias{filter_temporal}
\title{filter temporal}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
filter_temporal
}
\description{
filter temporal
}
\keyword{datasets}
