% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessGraph-class.R
\name{getPgidIndex}
\alias{getPgidIndex}
\title{Get Index of given process graph id in stored graphs}
\usage{
getPgidIndex(pgid)
}
\arguments{
\item{pgid}{process graph id}
}
\description{
Get Index of given process graph id in stored graphs
}
