% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{handleError}
\alias{handleError}
\title{Handle errors from process execution}
\usage{
handleError(e)
}
\arguments{
\item{e}{Error object containing the error message}
}
\value{
List with error code, message, and links
}
\description{
Handle errors from process execution
}
