% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutableProcess-class.R
\name{is.ExecutableProcess}
\alias{is.ExecutableProcess}
\title{Check if given process is a process}
\usage{
is.ExecutableProcess(obj)
}
\arguments{
\item{obj}{Process to be checked}
}
\description{
Check if given process is a process
}
