% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessGraph-class.R
\name{is.ProcessGraph}
\alias{is.ProcessGraph}
\title{Check if process graph}
\usage{
is.ProcessGraph(obj)
}
\arguments{
\item{obj}{process graph to be checked}
}
\description{
Check if process graph
}
