% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\name{Job}
\alias{Job}
\title{Job
This class represents an openEO job. Which is submitted by a user.}
\description{
Job
This class represents an openEO job. Which is submitted by a user.

Job
This class represents an openEO job. Which is submitted by a user.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Id of the job}

\item{\code{status}}{Current status of the job}

\item{\code{process}}{Process of the job}

\item{\code{created}}{When the job was created}

\item{\code{title}}{Title of the job}

\item{\code{description}}{Shortly description of the planned result}

\item{\code{results}}{Result of the executed process}

\item{\code{output}}{Format of the output}

\item{\code{openEoGraph}}{Store the openEo graph for the job info}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{output.folder}}{Set a new output folder}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Job-new}{\code{Job$new()}}
\item \href{#method-Job-setOutput}{\code{Job$setOutput()}}
\item \href{#method-Job-store}{\code{Job$store()}}
\item \href{#method-Job-load}{\code{Job$load()}}
\item \href{#method-Job-run}{\code{Job$run()}}
\item \href{#method-Job-jobInfo}{\code{Job$jobInfo()}}
\item \href{#method-Job-clone}{\code{Job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-new"></a>}}
\if{latex}{\out{\hypertarget{method-Job-new}{}}}
\subsection{Method \code{new()}}{
Initialize job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$new(id = NA, process = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of the job}

\item{\code{process}}{Process of the job}

\item{\code{status}}{Current status of the job}

\item{\code{created}}{When the job was created}

\item{\code{title}}{Title of the job}

\item{\code{description}}{Shortly description of the planned result}

\item{\code{results}}{Result of the executed process}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-setOutput"></a>}}
\if{latex}{\out{\hypertarget{method-Job-setOutput}{}}}
\subsection{Method \code{setOutput()}}{
Set the output format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$setOutput(output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output}}{Format of the output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-store"></a>}}
\if{latex}{\out{\hypertarget{method-Job-store}{}}}
\subsection{Method \code{store()}}{
Store the job in the Session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$store()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-load"></a>}}
\if{latex}{\out{\hypertarget{method-Job-load}{}}}
\subsection{Method \code{load()}}{
Load the job properties from the stored jobs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-run"></a>}}
\if{latex}{\out{\hypertarget{method-Job-run}{}}}
\subsection{Method \code{run()}}{
Execute the executable process graph and store it in the results of the job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The executed job
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-jobInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Job-jobInfo}{}}}
\subsection{Method \code{jobInfo()}}{
Get information about the job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$jobInfo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Info list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Job-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Job-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
