% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\docType{data}
\name{landsat_8_l1_c1}
\alias{landsat_8_l1_c1}
\title{landsat-8-l1-c1}
\format{
An object of class \code{Collection} (inherits from \code{R6}) of length 13.
}
\usage{
landsat_8_l1_c1
}
\description{
landsat-8-l1-c1
}
\keyword{datasets}
