% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{load_collection}
\alias{load_collection}
\title{load collection}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
load_collection
}
\description{
load collection
}
\keyword{datasets}
