% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_netCDF_cube.R
\docType{data}
\name{load_netCDF_cube}
\alias{load_netCDF_cube}
\title{load_netCDF_cube}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
load_netCDF_cube
}
\description{
load_netCDF_cube
}
\keyword{datasets}
