% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{load_process_metadata}
\alias{load_process_metadata}
\title{load_process_metadata}
\usage{
load_process_metadata(process_id)
}
\arguments{
\item{process_id}{The ID of the process to load metadata for}
}
\value{
A list containing the process metadata
}
\description{
Loads process metadata from a JSON file in the inst/processes directory
}
\keyword{internal}
