% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{load_stac}
\alias{load_stac}
\title{load stac}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
load_stac
}
\description{
load stac
}
\keyword{datasets}
