% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{merge_cubes}
\alias{merge_cubes}
\title{merge_cubes}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
merge_cubes
}
\description{
merge_cubes
}
\keyword{datasets}
