% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes-ml.R
\name{ml_datacube_schema}
\alias{ml_datacube_schema}
\title{Get Data Cube Schema}
\usage{
ml_datacube_schema()
}
\value{
A list containing:
\itemize{
\item{description}{A character string describing the data cube}
\item{schema}{A list defining the type and subtype of the data cube}
}
}
\description{
Returns a list containing the description and schema for a data cube object.
This schema is used to define the structure of data cubes in machine learning processes.
}
