% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes-ml.R
\docType{data}
\name{ml_fit}
\alias{ml_fit}
\title{Train an ML or DL model on extracted features}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
ml_fit
}
\arguments{
\item{model}{Object of subtype "mlm-model". The machine learning or deep learning model to train.}

\item{training_set}{Object of subtype "vector". Preprocessed training dataset containing features and target.}

\item{target}{Character. Name of the column in \code{training_set} that contains the target variable.}
}
\value{
A trained model object of subtype "mlm-model".
}
\description{
Trains a Machine Learning or Deep Learning model based on the provided
preprocessed training data and specified target variable.
}
\keyword{datasets}
