% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes-ml.R
\docType{data}
\name{mlm_class_lighttae}
\alias{mlm_class_lighttae}
\title{LightTAE for classification (experimental)}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
mlm_class_lighttae
}
\description{
Lightweight Temporal Self-Attention über die Zeitachse.
Eingabeform: (N, C, T). Pro Zeitstempel Linear-Projektion auf d_model,
dann scaled dot-product Self-Attention über T, Global-Pooling und
MLP-Head zur Klassifikation.
}
\keyword{datasets}
