% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes-ml.R
\docType{data}
\name{mlm_class_mlp}
\alias{mlm_class_mlp}
\title{MLP for classification}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
mlm_class_mlp
}
\value{
list(parameters=..., create_model=function(...){ nn_module })
}
\description{
Model case for a Multi-Layer Perceptron (MLP) classifier.
Defines hidden dense layers incl. activation + dropout, and returns
a factory function to create the torch nn_module.
}
\keyword{datasets}
