% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-processes.R
\docType{data}
\name{multiply}
\alias{multiply}
\title{multiply}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
multiply
}
\description{
multiply
}
\keyword{datasets}
