% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-processes.R
\docType{data}
\name{number_or_null_returns}
\alias{number_or_null_returns}
\title{Common returns for math processes}
\format{
An object of class \code{list} of length 2.
}
\usage{
number_or_null_returns
}
\description{
Common returns for math processes
}
\keyword{datasets}
