% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\docType{data}
\name{predict_model}
\alias{predict_model}
\title{predict_model}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
predict_model
}
\description{
predict_model
}
\keyword{datasets}
