% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-class.R
\name{Process}
\alias{Process}
\title{Process}
\description{
Process

Process
}
\details{
The process class in the openEO thinking. It is a described function that can be serialized in short and detail for the webservice,
but also relates to the internal R function (process) that will be applied to perform the task
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Id or name of the process}

\item{\code{summary}}{Shortly what the process does}

\item{\code{description}}{Extended description of the process}

\item{\code{categories}}{To which categories the process belongs}

\item{\code{parameters}}{Used parameters in the process}

\item{\code{returns}}{The output of the process}

\item{\code{operation}}{Function that executes the process}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Process-new}{\code{Process$new()}}
\item \href{#method-Process-processInfo}{\code{Process$processInfo()}}
\item \href{#method-Process-setParameter}{\code{Process$setParameter()}}
\item \href{#method-Process-clone}{\code{Process$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-new"></a>}}
\if{latex}{\out{\hypertarget{method-Process-new}{}}}
\subsection{Method \code{new()}}{
Initialize process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$new(
  id = NA,
  summary = NA,
  description = NA,
  categories = NA,
  parameters = NA,
  returns = NA,
  operation = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id or name of the process}

\item{\code{summary}}{Shortly what the process does}

\item{\code{description}}{Extended description of the process}

\item{\code{categories}}{To which categories the process belongs}

\item{\code{parameters}}{Used parameters in the process}

\item{\code{returns}}{The output of the process}

\item{\code{operation}}{Function that executes the process}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-processInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Process-processInfo}{}}}
\subsection{Method \code{processInfo()}}{
List information about the process
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$processInfo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-setParameter"></a>}}
\if{latex}{\out{\hypertarget{method-Process-setParameter}{}}}
\subsection{Method \code{setParameter()}}{
Set parameter for the executable process graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$setParameter(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of parameter}

\item{\code{value}}{Value to be set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Process-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Process-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Process$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
