% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{reduce_dimension}
\alias{reduce_dimension}
\title{reduce dimension}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
reduce_dimension
}
\description{
reduce dimension
}
\keyword{datasets}
