% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{rename_labels}
\alias{rename_labels}
\title{rename labels}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
rename_labels
}
\description{
rename labels
}
\keyword{datasets}
