% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{resample_spatial}
\alias{resample_spatial}
\title{resample spatial}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
resample_spatial
}
\description{
resample spatial
}
\keyword{datasets}
