% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\docType{data}
\name{save_result}
\alias{save_result}
\title{save result}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
save_result
}
\description{
save result
}
\keyword{datasets}
