% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{schema_format}
\alias{schema_format}
\title{schema_format}
\usage{
schema_format(type, subtype = NULL, items = NULL)
}
\arguments{
\item{type}{data type}

\item{subtype}{subtype of the data}
}
\value{
list with type and subtype(optional)
}
\description{
format for the schema
}
