% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\docType{data}
\name{sentinel_s2_l2a}
\alias{sentinel_s2_l2a}
\title{sentinel-s2-l2a}
\format{
An object of class \code{Collection} (inherits from \code{R6}) of length 13.
}
\usage{
sentinel_s2_l2a
}
\description{
sentinel-s2-l2a
}
\keyword{datasets}
