% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\docType{data}
\name{sentinel_s2_l2a_cogs}
\alias{sentinel_s2_l2a_cogs}
\title{sentinel-s2-l2a-cogs}
\format{
An object of class \code{Collection} (inherits from \code{R6}) of length 12.
}
\usage{
sentinel_s2_l2a_cogs
}
\description{
sentinel-s2-l2a-cogs
}
\keyword{datasets}
