% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SessionConfig-Class.R
\name{SessionConfig}
\alias{SessionConfig}
\title{SessionConfig class}
\usage{
SessionConfig(api.port = NULL, host = NULL)
}
\arguments{
\item{api.port}{On which port to run the plumer API}

\item{host}{Host of the plumber API}
}
\description{
Session configuration
}
