% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session-Class.R
\name{SessionInstance}
\alias{SessionInstance}
\title{Session class}
\description{
Session class

Session class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{graphs}}{User defined process graphs}

\item{\code{processes}}{Package processes}

\item{\code{data}}{Data of current Session}

\item{\code{jobs}}{Stored jobs of current Session}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SessionInstance-new}{\code{SessionInstance$new()}}
\item \href{#method-SessionInstance-getEndpoints}{\code{SessionInstance$getEndpoints()}}
\item \href{#method-SessionInstance-getConfig}{\code{SessionInstance$getConfig()}}
\item \href{#method-SessionInstance-getToken}{\code{SessionInstance$getToken()}}
\item \href{#method-SessionInstance-setToken}{\code{SessionInstance$setToken()}}
\item \href{#method-SessionInstance-startSession}{\code{SessionInstance$startSession()}}
\item \href{#method-SessionInstance-setBaseUrl}{\code{SessionInstance$setBaseUrl()}}
\item \href{#method-SessionInstance-getBaseUrl}{\code{SessionInstance$getBaseUrl()}}
\item \href{#method-SessionInstance-initDirectory}{\code{SessionInstance$initDirectory()}}
\item \href{#method-SessionInstance-initEndpoints}{\code{SessionInstance$initEndpoints()}}
\item \href{#method-SessionInstance-createEndpoint}{\code{SessionInstance$createEndpoint()}}
\item \href{#method-SessionInstance-assignData}{\code{SessionInstance$assignData()}}
\item \href{#method-SessionInstance-assignProcess}{\code{SessionInstance$assignProcess()}}
\item \href{#method-SessionInstance-assignJob}{\code{SessionInstance$assignJob()}}
\item \href{#method-SessionInstance-runJob}{\code{SessionInstance$runJob()}}
\item \href{#method-SessionInstance-clone}{\code{SessionInstance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-new"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-new}{}}}
\subsection{Method \code{new()}}{
Create a new session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$new(configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configuration}}{Session configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-getEndpoints"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-getEndpoints}{}}}
\subsection{Method \code{getEndpoints()}}{
Get endpoints
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$getEndpoints()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
endpoints
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-getConfig"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-getConfig}{}}}
\subsection{Method \code{getConfig()}}{
Get configuration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$getConfig()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
configuration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-getToken"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-getToken}{}}}
\subsection{Method \code{getToken()}}{
Get token
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$getToken()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-setToken"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-setToken}{}}}
\subsection{Method \code{setToken()}}{
Assign this token to the session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$setToken(new)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new}}{New token}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-startSession"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-startSession}{}}}
\subsection{Method \code{startSession()}}{
Start the session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$startSession()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-setBaseUrl"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-setBaseUrl}{}}}
\subsection{Method \code{setBaseUrl()}}{
Set base url
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$setBaseUrl(port, host)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{port}}{Which port is currently saved}

\item{\code{host}}{Which host is currently saved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-getBaseUrl"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-getBaseUrl}{}}}
\subsection{Method \code{getBaseUrl()}}{
Get base url
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$getBaseUrl()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
base url
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-initDirectory"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-initDirectory}{}}}
\subsection{Method \code{initDirectory()}}{
initializes workspace and data paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$initDirectory()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-initEndpoints"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-initEndpoints}{}}}
\subsection{Method \code{initEndpoints()}}{
build a df to add the endpoints later on
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$initEndpoints()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-createEndpoint"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-createEndpoint}{}}}
\subsection{Method \code{createEndpoint()}}{
Create an endpoint
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$createEndpoint(
  path,
  method,
  handler = NULL,
  filter = FALSE,
  serializer = serializer_unboxed_json()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path for the endpoint}

\item{\code{method}}{type of request}

\item{\code{handler}}{function to be executed}

\item{\code{filter}}{deactivate filter for several endpoints}

\item{\code{serializer}}{plumber serializer to be used}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
created Endpoint
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-assignData"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-assignData}{}}}
\subsection{Method \code{assignData()}}{
Function to assign data of collection to the data path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$assignData(col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{Collection of class 'Collection'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-assignProcess"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-assignProcess}{}}}
\subsection{Method \code{assignProcess()}}{
Function to assign the process to the Session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$assignProcess(pro)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pro}}{Process of class 'Process'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-assignJob"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-assignJob}{}}}
\subsection{Method \code{assignJob()}}{
Function to assign a job to the Session
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$assignJob(job)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job}}{Job of class 'Job'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-runJob"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-runJob}{}}}
\subsection{Method \code{runJob()}}{
Execute the job
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$runJob(job)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{job}}{Job to be executed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionInstance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SessionInstance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionInstance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
