% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{stac_call}
\alias{stac_call}
\title{stacCall}
\usage{
stac_call(id, spatial_extent, temporal_extent)
}
\arguments{
\item{id}{Collection ID}

\item{spatial_extent}{bbox of the region of interest}

\item{temporal_extent}{Time range}
}
\description{
Internal function to perform stac api calls
}
