% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{throwError}
\alias{throwError}
\title{Throw an OpenEO-compliant error}
\usage{
throwError(id, ...)
}
\arguments{
\item{id}{Error identifier (e.g., "JobNotFound", "Internal")}

\item{...}{Named variables to substitute in the error message template}
}
\value{
Stops execution with a JSON-formatted error message
}
\description{
Throw an OpenEO-compliant error
}
\details{
This function looks up the error definition by ID and formats it according to
OpenEO error specifications. Variable substitution is performed on the message
template using the provided named arguments.

Error structure:
\itemize{
\item code: OpenEO-specific error code
\item msg: Human-readable error message (with variable substitution)
\item status: HTTP status code to be returned
}
}
