% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\docType{data}
\name{train_model}
\alias{train_model}
\title{train_model}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
train_model
}
\description{
train_model
}
\keyword{datasets}
