% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes-cube.R
\docType{data}
\name{vec_datacube}
\alias{vec_datacube}
\title{return object for the processes}
\format{
An object of class \code{list} of length 2.
}
\usage{
vec_datacube
}
\description{
return object for the processes
}
\keyword{datasets}
