% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\name{writeJobInfo}
\alias{writeJobInfo}
\title{Write the job properties to a txt file}
\usage{
writeJobInfo(job)
}
\arguments{
\item{job}{Job from which the properties will be stored}
}
\description{
Write the job properties to a txt file
}
