% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\name{Collection}
\alias{Collection}
\title{Collection}
\description{
Collection

Collection
}
\details{
This class represents the collections, which contain a set of Granules.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Id of the collection}

\item{\code{title}}{Collection title}

\item{\code{description}}{Short description of the collection}

\item{\code{spatialExtent}}{Spatial extent of the collection}

\item{\code{temporalExtent}}{Temporal extent of the collection}

\item{\code{bands}}{Bands of the collection}

\item{\code{constellation}}{Constellation of the collection}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Collection-new}{\code{Collection$new()}}
\item \href{#method-Collection-collectionInfo}{\code{Collection$collectionInfo()}}
\item \href{#method-Collection-collectionInfoExtended}{\code{Collection$collectionInfoExtended()}}
\item \href{#method-Collection-clone}{\code{Collection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-new"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-new}{}}}
\subsection{Method \code{new()}}{
Initialize collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$new(
  id = NA,
  title = NA,
  description = NA,
  spatialExtent = NA,
  temporalExtent = NA,
  bands = NA,
  constellation = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Id of the collection}

\item{\code{title}}{Collection title}

\item{\code{description}}{Short description of the collection}

\item{\code{spatialExtent}}{Spatial extent of the collection}

\item{\code{temporalExtent}}{Temporal extent of the collection}

\item{\code{bands}}{Bands of the collection}

\item{\code{constellation}}{Constellation of the collection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-collectionInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-collectionInfo}{}}}
\subsection{Method \code{collectionInfo()}}{
Get collection information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$collectionInfo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing collection information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-collectionInfoExtended"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-collectionInfoExtended}{}}}
\subsection{Method \code{collectionInfoExtended()}}{
Get extended collection information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$collectionInfoExtended()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing extended collection information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Collection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Collection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Collection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
