% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes-ml.R
\docType{data}
\name{ml_predict}
\alias{ml_predict}
\title{Apply a trained ML/DL model to a data cube and return predictions}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
ml_predict
}
\arguments{
\item{cube}{Object of subtype "raster-cube". The area-of-interest data cube on which predictions will be made.}

\item{model}{Object of subtype "mlm-model" or a file path (ONNX, .rds, or .pt). The trained model to use for prediction.}
}
\value{
A raster-cube object containing the predicted values in a new "prediction" band.
}
\description{
Applies a machine learning or deep learning model to every pixel (and/or time step)
of the input raster-cube and returns a new data cube with the predicted values.
}
\keyword{datasets}
