% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-processes.R
\docType{data}
\name{ml_predict}
\alias{ml_predict}
\title{TO DO : Implement machine learning processes}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
ml_predict
}
\description{
ml_predict
}
\keyword{datasets}
