% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-processes.R
\docType{data}
\name{save_ml_model}
\alias{save_ml_model}
\title{save_ml_model}
\format{
An object of class \code{Process} (inherits from \code{R6}) of length 12.
}
\usage{
save_ml_model
}
\description{
save_ml_model
}
\keyword{datasets}
